requirements.txt
pandas==1.5.1
psycopg2-binary==2.9.5
requests==2.26.0
